+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ﻿## [Unreleased] ##
 ### Fixed ###
 
+  * [retail] The dungeon ready popup had two backdrops
   * [retail] The world map's close button was out of place when maximized
   * [retail] Intermittent errors when opening the talent frame